fixed bug that caused int overflow on 32bit systems

+++ frame.go
@@ -364,7 +364,7 @@ func NewFrame(tm time.Time, evtnum Evtnum, v0 float64, v1 int64, data []byte) (*
 		return nil, EvtnumOutOfRangeErr
 	}
 
-	if len(data) > (1<<43)-1 {
+	if int64(len(data)) > (1<<43)-1 {
 		return nil, DataTooBigErr
 	}
